+++ larq_compute_engine/tf/tests/BUILD
@@ -9,7 +9,7 @@ cc_test(
     copts = ["-Iext/gtest/include"],
     deps = [
         "@com_google_googletest//:gtest_main",
-        "//larq_compute_engine/core:bgemm",
+        "//larq_compute_engine/core:bgemm_functor",
     ],
 )
 
@@ -32,7 +32,7 @@ cc_test(
     deps = [
          "@com_google_googletest//:gtest_main",
          "@eigen_archive//:eigen",
-         "//larq_compute_engine/core:fused_bgemm",
+         "//larq_compute_engine/core:fused_bgemm_functor",
          "//larq_compute_engine/core:utils",
     ],
 )
@@ -44,7 +44,7 @@ cc_test(
     copts = ["-Iext/gtest/include"],
     deps = [
          "@com_google_googletest//:gtest_main",
-         "//larq_compute_engine/core:im2col",
+         "//larq_compute_engine/core:im2col_functor",
     ],
 )
 
@@ -55,7 +55,7 @@ cc_test(
     copts = ["-Iexternal/gtest/include"],
     deps = [
          "@com_google_googletest//:gtest_main",
-         "//larq_compute_engine/core:bconv2d",
+         "//larq_compute_engine/core:bconv2d_functor",
     ],
 )
